refactor: 케밥 네이밍으로 변경 (#97)
