+++ api/routes/auth/signin/index.ts
@@ -2,8 +2,8 @@ import express, { Request, Response } from 'express';
 
 import authService from 'services/auth';
 
-import errorGenerator from 'utils/errorGenerator';
-import errorHandler from 'utils/errorHandler';
+import errorGenerator from 'utils/error-generator';
+import errorHandler from 'utils/error-handler';
 
 const router = express.Router();
