+++ api/services/auth.ts
@@ -1,7 +1,7 @@
 import { VerifyErrors } from 'jsonwebtoken';
 
 import sequelize, { db } from 'models/db';
-import errorGenerator from 'utils/errorGenerator';
+import errorGenerator from 'utils/error-generator';
 import { createToken, verifyToken } from 'utils/jwt';
 import { hashPassword, checkPassword } from 'utils/encryption';
